refactor: move tools in one dir
