+++ example/index.ts
@@ -3,8 +3,8 @@ import { DummyLLM, OpenAILLM, LLM } from "../src/llm";
 import {
   CurrentWeatherAPITool,
   ForecastWeatherAPITool,
-} from "../src/tools/weather/nubila";
-import { NewsAPITool } from "../src/tools/news/newsapi";
+} from "../src/tools/nubila";
+import { NewsAPITool } from "../src/tools/newsapi";
 import { Agent } from "../src/agent";
 import { Tool } from "../src/tools/tool";
 import * as dotenv from "dotenv";
