+++ src/tools/dify.ts
@@ -1,9 +1,9 @@
 import axios from "axios";
 
-import { APITool } from "../tool";
-import { handleStreamResponse } from "../../utils/stream_utils";
-import { extractContentFromTags } from "../../utils/parsers";
-import { LLMService } from "../../services/llm-service";
+import { APITool } from "./tool";
+import { handleStreamResponse } from "../utils/stream_utils";
+import { extractContentFromTags } from "../utils/parsers";
+import { LLMService } from "../services/llm-service";
 
 export class DePINTool extends APITool<any> {
   constructor() {
